Updated physics, so to work more like 60fps.
git-svn-id: svn://10.0.50.55/puffbomb/trunk@2443 d14347b8-48fd-0310-9e84-97907d434d85

+++ Code/Game/PuffBOMB/GolfGame/GolfGameEngine.cpp
@@ -696,7 +696,7 @@ void cGolfGameEngine::TurnBasedPlay() {
 					Engine2D::cImpulse MyImpulse(
 							BombPos,
 							// Inner Radius, Intensity, Tangent //
-							Real( 0 ), Real( 30 ), Real( 0 ),
+							Real( 0 ), Real( 18 ), Real( 0 ),
 							// Outer Radius, Intensity, Tangent //
 							Real( 512 ), Real( 0 ), Real( 0 )
 							);
